Refactor device flashing logic
