+++ gtk/src/app/mod.rs
@@ -8,20 +8,25 @@ use self::content::Content;
 pub use self::dialogs::OpenDialog;
 use self::header::Header;
 pub use self::misc::*;
-pub use self::state::{Connect, State};
+pub use self::state::{Connect, FlashTask, State};
 
 // TODO: Use AtomicU64 / Bool when https://github.com/rust-lang/rust/issues/32976 is stable.
 
-use std::path::PathBuf;
+use std::mem;
+use std::path::{Path, PathBuf};
 use std::process;
 use std::rc::Rc;
-use std::sync::Arc;
+use std::sync::{Arc, Mutex};
+use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::mpsc::{Sender, Receiver};
+use std::thread;
+use std::time::{Duration, Instant};
 use std::fs::File;
 use std::thread::JoinHandle;
 
+use block::BlockDevice;
 use flash::FlashRequest;
-use popsicle::mnt::MountEntry;
+use popsicle::mnt::{self, MountEntry};
 use popsicle::{self, DiskError};
 
 use gtk;
@@ -97,6 +102,177 @@ impl AppWidgets {
         {
             self.set_error(state, &why);
         }
+    }
+
+    pub fn switch_to_device_flashing(&self, state: &State) {
+        let back = &self.header.back;
+        let next = &self.header.next;
+        let stack = &self.content.container;
+        let summary_grid = &self.content.flash_view.progress_list;
+        let task_handles = &state.task_handles;
+        let bars = &state.bars;
+        let start = &state.start;
+        let tasks = &state.tasks;
+
+        macro_rules! try_or_error {
+            ($action:expr, $msg:expr) => {{
+                match $action {
+                    Ok(value) => value,
+                    Err(why) => {
+                        self.set_error(state, &format!("{}: {}", $msg, why));
+                        return;
+                    }
+                }
+            }}
+        }
+
+        // Wait for the flash state to be 0 before proceeding.
+        while state.flash_state.load(Ordering::SeqCst) != 0 {
+            thread::sleep(Duration::from_millis(16));
+        }
+
+        let mut data = Vec::new();
+        let image_data: Arc<Vec<u8>>;
+
+        {
+            let mut image_buffer_lock = try_or_error!(
+                state.buffer.data.lock(),
+                "failed to lock buffer.data mutex"
+            );
+
+            let device_list = try_or_error!(
+                state.devices.lock(),
+                "device list mutex lock failure"
+            );
+
+            let devs = device_list
+                .iter()
+                .filter(|x| x.1.get_active())
+                .map(|x| x.0.clone())
+                .collect::<Vec<_>>();
+
+            let mounts = try_or_error!(
+                mnt::get_submounts(Path::new("/")),
+                "unable to obtain mount points"
+            );
+
+            try_or_error!(
+                state.devices_request.send((devs, mounts.clone())),
+                "unable to send device request"
+            );
+
+            let disks_result = try_or_error!(
+                state.devices_response.recv(),
+                "unable to get device request response"
+            );
+
+            let disks = try_or_error!(
+                disks_result,
+                "unable to get devices"
+            );
+
+            back.get_style_context().map(|c| {
+                c.remove_class("back-button");
+                c.add_class("destructive-action");
+            });
+
+            back.set_label("Cancel");
+            back.set_visible(true);
+            next.set_visible(false);
+            stack.set_visible_child_name("flash");
+
+            // Clear the progress bar summaries.
+            let mut bars = bars.borrow_mut();
+            bars.clear();
+            summary_grid.get_children().iter().for_each(|c| c.destroy());
+
+            *start.borrow_mut() = Instant::now();
+            let mut tasks = try_or_error!(
+                tasks.lock(),
+                "tasks mutex lock failure"
+            );
+
+            let mut task_handles = try_or_error!(
+                task_handles.lock(),
+                "task handles mutex lock failure"
+            );
+
+            state.flash_state.store(1, Ordering::SeqCst);
+
+            // Take ownership of the data, so that we may wrap it within an `Arc`
+            // and redistribute it across threads.
+            //
+            // Note: Possible optimization could be done to avoid the wrap.
+            //       Avoiding the wrap could eliminate two allocations.
+            image_data = {
+                let (_, ref mut image_data) = *image_buffer_lock;
+                mem::swap(&mut data, image_data);
+                Arc::new(data)
+            };
+
+            for (id, (disk_path, mut disk)) in disks.into_iter().enumerate() {
+                let id = id as i32;
+                let image_data = image_data.clone();
+                let progress = Arc::new(AtomicUsize::new(0));
+                let finished = Arc::new(AtomicUsize::new(0));
+                let pbar = ProgressBar::new();
+                pbar.set_hexpand(true);
+
+                let label = {
+                    let disk_path = try_or_error!(
+                        Path::new(&disk_path).canonicalize(),
+                        format!("unable to get canonical path of {}", disk_path)
+                    );
+                    if let Some(block) = BlockDevice::new(&disk_path) {
+                        gtk::Label::new(
+                            [&block.label(), " (", &disk_path.to_string_lossy(), ")"]
+                                .concat()
+                                .as_str(),
+                        )
+                    } else {
+                        gtk::Label::new(disk_path.to_string_lossy().as_ref())
+                    }
+                };
+
+                label.set_justify(gtk::Justification::Right);
+                label.get_style_context().map(|c| c.add_class("bold"));
+                let bar_label = gtk::Label::new("");
+                bar_label.set_halign(gtk::Align::Center);
+                let bar_container = gtk::Box::new(Orientation::Vertical, 0);
+                bar_container.pack_start(&pbar, false, false, 0);
+                bar_container.pack_start(&bar_label, false, false, 0);
+                summary_grid.attach(&label, 0, id, 1, 1);
+                summary_grid.attach(&bar_container, 1, id, 1, 1);
+                bars.push((pbar, bar_label));
+
+                // Spawn a thread that will update the progress value over time.
+                //
+                // The value will be stored within an intermediary atomic integer,
+                // because it is unsafe to send GTK widgets across threads.
+                task_handles.push({
+                    let _ = state.flash_request.send(FlashRequest::new(
+                        disk,
+                        disk_path,
+                        image_data.len() as u64,
+                        image_data,
+                        state.flash_state.clone(),
+                        progress.clone(),
+                        finished.clone()
+                    ));
+
+                    state.flash_response.recv().expect("expected join handle to be returned")
+                });
+
+                tasks.push(FlashTask {
+                    previous: Arc::new(Mutex::new([0; 7])),
+                    progress,
+                    finished,
+                });
+            }
+        }
+
+        state.async_reattain_image_data(image_data);
+        summary_grid.show_all();
     }
 
     pub fn set_error(&self, state: &State, msg: &str) {
